fix: libicu version lookup failed for security releases
The libicu version lookup failed if the distribution in the changelog
file is *-security, which is the case for security releases.

This commit fixes this issue by checking if the series is contained
in the distribution string rather than checking for equality.

NOTE: this commit also adds the lookup values for kinetic and lunar,
which are both obsolete. I just added these values, because we once
built packages for these series and it does not take up a lot of space.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>