fix: use env substitution variables
Some dh_* tools do not substitute the values for the variables defined
in debian/substvars, but they can substiture the values from environment
variables.

This bug was observed for: dh_installdirs and dh_install.

Accoring to the documentation they should support subsitution from
debian/substvars with compat level 13, which is defined in
debian/control.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>