fix: set DOTNET_80_OR_GREATER, etc. correctly
The variables DOTNET_60_OR_GREATER, DOTNET_70_OR_GREATER and
DOTNET_80_OR_GREATER are not set correctly. This commit sets the
correct values.

This issue lead to dotnet6 and dotnet7 packages failing to build,
because commands for DOTNET_80_OR_GREATER were executed which are by
definition not suitable for dotnet6 and dotnet7.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>