fix: shbang of debian/rules
The shbang of a debian/rules file is supposed to be `#!/usr/bin/make -f`
and therefore was flagged by lintian as an error:

```
E: debian-rules-not-a-makefile
N:
N:   The debian/rules file for this package does not appear to be a makefile or does not start with the required line. debian/rules must be a valid makefile and must have "#!/usr/bin/make -f" as its first line.
N:
N:   Please refer to Main building script: debian/rules (Section 4.9) in the Debian Policy Manual for details.
N:
```