fix: lintian warnings executable-not-elf-or-script
Fixes the lintian warnings executable-not-elf-or-script for the
`dotnet-sdk-aot-9.0` package.

```
W: executable-not-elf-or-script
N:
N:   This executable file is not an ELF format binary, and does not start with the #! sequence that marks interpreted scripts. It might be a sh script that fails to name /bin/sh as its shell, or it may be incorrectly marked as executable. Sometimes upstream files developed on Windows are marked unnecessarily as executable on
N:   other systems.
N:
N:   If you are using debhelper to build your package, running dh_fixperms will often correct this problem for you.
N:
N:   Please refer to Scripts (Section 10.4) in the Debian Policy Manual for details.
```