chore: remove dotnet6, dotnet7 & mantic files
This commit removes files for .NET versions and
target platforms that will never be deployed.

- .NET 7 went End of Life on May 14, 2024 [1]
- .NET 6 will go End of Life on November 12, 2024 [2]
- Ubuntu 23.10 (Mantic Minotaur) went End of Life
  on July 11, 2024 [2]

It makes no sense in maintaining related files,
when they are effectively unused. The original
plan was to deploy this project to these platforms,
but 1. I did not pay enough attention to the EoL
dates and 2. this effort took longer than expected.

[1] https://dotnet.microsoft.com/en-us/platform/support/policy/dotnet-core
[2] https://lists.ubuntu.com/archives/ubuntu-security-announce/2024-July/008463.html

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>