build: add +build1 to version of runtime
This is needed, because the December 2024 release only incremented the
SDK version identifier, but not the runtime version identifier.
This would result in a conflict in the archive, because two source
packages would procuce binary packages with the same version number.
FO127 mandates that a `+build1` is added to the runtime version to solve
this conflict.