fix(lintian,net8): suppress manpage troff warnings
Troff complains that it is silly that the dotnet8 manpages select a
monospace font on a terminal output that only supports monospace fonts.

We may voice this upstream, but the output is generated by pandoc and
there is no user/developer impact. We will silence this warning as it
is effectively just noise to more important warnings.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>