fix(changelog): upstream updated release info
Microsoft updated the release info for the January 2025 .NET release.
It was reclassified from a non-security release to a security release.

The following versions were already deployed:
- 9.0.102-9.0.1-0ubuntu1 (dotnet9/plucky)
- 9.0.102-9.0.1-0ubuntu1~24.04.1~ppa1 (dotnet9/noble)
- 9.0.102-9.0.1-0ubuntu1~22.04.1~ppa1 (dotnet9/jammy)
- 8.0.112-8.0.12-0ubuntu2 (dotnet8/plucky)

Therefore these versions received a no-change rebuild with the included
security infos and deployed to the -security pocket.

The following versions were still in the queue:
- 9.0.102-9.0.1-0ubuntu1~24.10.1 (dotnet9/oracular)
- 8.0.112-8.0.12-0ubuntu1~24.10.1 (dotnet8/oracular)
- 8.0.112-8.0.12-0ubuntu1~24.04.1 (dotnet8/noble)
- 8.0.112-8.0.12-0ubuntu1~22.04.1 (dotnet8/jammy)

Therefore these versions were rejected from the upload queue and
re-uploaded with the same version number and the correct changelog entry.
Deployed to -security.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>