fix: moved Suggests & Recommends to corresponding binary packages
The Recommends `dotnet-sdk-aot-*` and Suggests `dotnet-sdk-dbg-*`,
`dotnet-runtime-dbg-*`, `aspnetcore-runtime-dbg-*` were part of the
`dotnet*` virtual package.

This is a minor issue, because these Suggests and Recommends do not get
installed when just installing the runtime or when someone installs the
`dotnet-sdk-*` package directly instead of using the `dotnet*` package.

This commit moved
- `Recommends: dotnet-sdk-aot-* [amd64 arm64]` to the `dotnet-sdk-*` package
- `Suggests: dotnet-sdk-dbg-*` to the `dotnet-sdk-*` package
- `Suggests: dotnet-runtime-dbg-*` to the `dotnet-runtime-*` package
- `Suggests: aspnetcore-runtime-dbg-*` to the `aspnetcore-runtime-dbg-*` package

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>