fix: disable strict artifact RID matching
No longer require that the .NET artifacts used for
bootstrapping the .NET SDK have the same RID as the
build output.

This strict matching caused issues during bootstrapping
of .NET for a new Ubuntu series, because it was build with
the binary artifact of the previous series, which caused
the RIDs not to match.

Fixes: https://github.com/canonical/dotnet-source-build/issues/12
Co-authored-by: Mateus Rodrigues de Morais <mateus.morais@canonical.com>
Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>