fix: pass through HOME when invoking prep script
When we invoke the prep script in the build-orig-tarball script we have
to pass through the HOME environment variable.

This was done implicitly before the DOTNET_NOLOGO and
DOTNET_SKIP_FIRST_TIME_EXPERIENCE were set that changed the behaviour
of the `subprocess.run` call:

> If env is not None, it must be a mapping that defines the environment
> variables for the new process; these are used instead of the default
> behavior of inheriting the current process’ environment.

This commit also creates a temporary home if none exists in the parent
environment.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>