feat: bulk changelog entry creation maintainer tool
Adds a maintainer tool that appends a changelog entry
for a new upstream release for all distributions of a
source package.

Rationale:
This tool reduces the likelihood of copy&paste errors
when creating the various changelog entries.

Note:
The current state of the tool is very basic. It's
functionality may be extended in the future.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>